pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.17.0",
    "sse-starlette>=0.9.0",
    "dependency-injector>=4.48.2",
    "fastapi[standard]>=0.118.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.0.1",
    "ujson>=5.11.0",
    "requests>=2.32.5",
    "prometheus-client>=0.21.0",
    "psutil>=6.1.0",
    "aiohttp>=3.13.2",
    "asyncmy>=0.2.10",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "sse-starlette",
]

[dependency-groups]
dev = [
    # Development tools
    "black>=24.10.0",
    "flake8>=7.1.0",
    "isort>=5.13.0",
    "pre-commit>=4.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "safety>=3.2.0",
    "bandit>=1.8.0",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
asyncio_mode = "auto"
